Preliminary preparation steps
This is branch intended to replace the reinvented wheel of the manual
panzoom implementation that doesn't work for touch and does not have
inertia.

I realised that panzoom does not work as intended as it allows the
entire element to move around, instead of the content inside the canvas.
Additionally, it probably does not cooperate well touch events.
I don't know a good library that:
    - is lightweight
    - works for touch
    - has inertia
    - prevents default actions except for click