Prevented canvas scroll previous page & highlight.
When swiping right quickly on Chrome the previous page action was
triggered.

There is a blue highlight that occurs only a mobile and can be
reproduced by using the mobile mode on the inspector. It is the same
feature that makes links go blue when clicked, however, it makes the
entire canvas flash not just the relevant icon. This could be
implemented manually if desired, and flash red if too expensive.